Add files via upload
HR Analytics Dashboard using Microsoft Power BI. In this project, the company wants to know about Employee Demographics where they want to check the analytics that, what is the reason for the majority attrition rate in the organization. So that they can analyze the root cause and resolve it according to the reasons they get from the Dashboard.