Add files via upload
Power BI Project on Presence Rate of Employees of Atliq Software Company. HR of the company wants to understand the Employee's preferences for their working places. What are the Sick Leave Rate, WFH Percentage as well as the Presence Rate?  So that they can plan their team meetings in a better way and also serve clients when they are at full potential.