Add files via upload
Tableau Project: Financial Complaint Dashboard, In this project, we have to define what are the problems or customer complaints and their resolution. For which product is the complaint rate highest or lowest? What are the issues regarding the usage of the product? What is the Percentage Dispute of the Customers and resolved?
Complete Interactive Project. Also published in Tableau Public.